broke out react files to make more modular and set up some component directories.
