+++ src/components/GroupGuru.css
@@ -1,0 +1,7 @@
+* {
+  padding: 0px;
+  margin: 0px;
+}
+#root {
+  background-color: #282c34;
+}
