+++ src/components/GroupGuru.js
@@ -1,0 +1,18 @@
+import React, { Component } from "react"
+import NavBar from "./nav/NavBar"
+import ApplicationViews from "./ApplicationViews"
+import "./GroupGuru.css";
+import "bootstrap/dist/css/bootstrap.min.css"
+
+class GroupGuru extends Component {
+  render() {
+      return (
+          <React.Fragment>
+              <NavBar />
+              <ApplicationViews />
+          </React.Fragment>
+      )
+  }
+}
+
+export default GroupGuru
